pyproject.toml
[project]
name = "facefinder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Joseph Chazalon", email = "joseph.chazalon@gmail.com" }
]
requires-python = ">=3.9"
dependencies = [
    "facenet-pytorch>=2.6.0",
    "faiss-cpu>=1.12.0",
    "jinja2>=3.1.6",
    "mtcnn>=0.1.1",
    "numpy>=1.26.4",
    "pillow>=10.2.0",
    "tensorflow-cpu>=2.18.0",
    "torch>=2.2.2",
    "tqdm>=4.67.1",
]

[project.scripts]
facefinder = "facefinder:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
